Implement Features : Add Consts
